style: merge files
