+++ src/main.jsx
@@ -2,7 +2,7 @@ import { createRoot } from "react-dom/client";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import { StrictMode } from "react";
 import App from "./App.jsx";
-import ProductChat from "./pages/ProductChat/index.jsx";
+import ProductChat from "./pages/ProductChat";
 import "./index.css";
 
 createRoot(document.getElementById("root")).render(
