--- src/pages/ProductChat/index.jsx
@@ -1,15 +1,0 @@
-import Header from "./Header";
-import Chat from "./Chat";
-import TypeIn from "./TypeIn";
-
-function ProductChat() {
-  return (
-    <div className=" font-TC w-[375px] mx-auto">
-      <Header />
-      <Chat />
-      <TypeIn />
-    </div>
-  );
-}
-
-export default ProductChat;
